+++ actors/migration/nv10/power.go
@@ -1,4 +1,4 @@
-package nv9
+package nv10
 
 import (
 	"context"
