+++ actors/migration/nv10/test/parallel_migration_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"golang.org/x/sync/errgroup"
 
-	"github.com/filecoin-project/specs-actors/v3/actors/migration/nv9"
+	"github.com/filecoin-project/specs-actors/v3/actors/migration/nv10"
 )
 
 func TestParallelMigrationCalls(t *testing.T) {
@@ -27,7 +27,7 @@ func TestParallelMigrationCalls(t *testing.T) {
 	// Run migration
 	adtStore := adt2.WrapStore(ctx, cbor.NewCborStore(bs))
 	startRoot := vm.StateRoot()
-	endRootSerial, err := nv9.MigrateStateTree(ctx, adtStore, startRoot, abi.ChainEpoch(0), nv9.Config{MaxWorkers: 1}, log, nv9.NewMemMigrationCache())
+	endRootSerial, err := nv10.MigrateStateTree(ctx, adtStore, startRoot, abi.ChainEpoch(0), nv10.Config{MaxWorkers: 1}, log, nv10.NewMemMigrationCache())
 	require.NoError(t, err)
 
 	// Migrate in parallel
@@ -35,12 +35,12 @@ func TestParallelMigrationCalls(t *testing.T) {
 	grp, ctx := errgroup.WithContext(ctx)
 	grp.Go(func() error {
 		var err1 error
-		endRootParallel1, err1 = nv9.MigrateStateTree(ctx, adtStore, startRoot, abi.ChainEpoch(0), nv9.Config{MaxWorkers: 2}, log, nv9.NewMemMigrationCache())
+		endRootParallel1, err1 = nv10.MigrateStateTree(ctx, adtStore, startRoot, abi.ChainEpoch(0), nv10.Config{MaxWorkers: 2}, log, nv10.NewMemMigrationCache())
 		return err1
 	})
 	grp.Go(func() error {
 		var err2 error
-		endRootParallel2, err2 = nv9.MigrateStateTree(ctx, adtStore, startRoot, abi.ChainEpoch(0), nv9.Config{MaxWorkers: 2}, log, nv9.NewMemMigrationCache())
+		endRootParallel2, err2 = nv10.MigrateStateTree(ctx, adtStore, startRoot, abi.ChainEpoch(0), nv10.Config{MaxWorkers: 2}, log, nv10.NewMemMigrationCache())
 		return err2
 	})
 	require.NoError(t, grp.Wait())
